Open additional port(6444) for apiserver VIP
There is a corner case bug where some clients are connecting
directly to the $nodeIP:6443 and not using the apiserver
VIP with a long lived connection. If a new connection is
sent from the same client IP to the VIP and it is LB'd to
the node with the long lived connection, DNAT will modify
the VIP to be $nodeIP. If that happens AND the source tcp
port is the same on the two connections we have a collision
and the second connection is ignored and eventually times
out.

This commit will use socat to listen on port 6444 to
also handle traffic for 6443. The DNAT -j REDIRECT rule
will now also modify the VIP to $nodeIP, but the port
will be changed to be 6444 which will ensure there is no
conflict with any long lived connection directly to the
$nodeIP:6443

additional info:
https://docs.google.com/presentation/d/1p9nuhAIY0unlceiMvfosmVIVS4BOUDgVUtEdlasSJDw/edit#slide=id.g88fa8d5926_0_0
https://bugzilla.netfilter.org/show_bug.cgi?id=1427
https://github.com/kubernetes/kubernetes/issues/90258
https://gist.github.com/aojea/4ffb520db9f72de21685294d8fec7f33

Signed-off-by: Jamo Luhrsen <jluhrsen@gmail.com>